# specs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: OACP Observability API
  version: 1.0.0
  description: |
    Frozen `/v1/observability/*` surface for OACP v1.0 (Day 54).

    Breaking changes to these paths or response schemas require bumping `info.version`
    and updating `specs/openapi/v1.lock.json` (enforced in CI).

    Legacy `GET /playground/snapshot` is deprecated — use `GET /v1/observability/snapshot`.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: OACP
    url: https://github.com/naaa-G/OACP

servers:
  - url: http://127.0.0.1:3000
    description: Local reference server

tags:
  - name: observability
    description: Console and MCPLab observability surface

paths:
  /v1/openapi.json:
    get:
      tags: [observability]
      summary: OpenAPI document for `/v1/*` observability routes
      operationId: getOpenApiDocument
      responses:
        '200':
          description: OpenAPI 3.1 JSON document
          content:
            application/json:
              schema:
                type: object

  /v1/observability/runtime-config:
    get:
      tags: [observability]
      summary: Console runtime auth configuration (public)
      operationId: getObservabilityRuntimeConfig
      responses:
        '200':
          description: Auth hints for Console SSE and snapshot polling
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeConfigResponse'

  /v1/observability/snapshot:
    get:
      tags: [observability]
      summary: Unified observability snapshot for Console polling
      operationId: getObservabilitySnapshot
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: trace_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
          description: Focus trace for `active_trace` and graph context
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 25
          description: Max traces in `traces` list
      responses:
        '200':
          description: Snapshot payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SnapshotResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

  /v1/observability/traces/{traceId}/graph:
    get:
      tags: [observability]
      summary: Trace-scoped delegation graph for Ops 2D mode
      operationId: getObservabilityTraceGraph
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      parameters:
        - name: traceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Trace graph nodes and edges
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceGraphResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /v1/observability/events:
    get:
      tags: [observability]
      summary: Server-sent events stream for live Console updates
      operationId: streamObservabilityEvents
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
        - SseApiKeyQuery: []
      parameters:
        - name: trace_id
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: after
          in: query
          required: false
          schema:
            type: string
          description: Resume after event id (alias for Last-Event-ID)
        - name: api_key
          in: query
          required: false
          schema:
            type: string
          description: API key for browser EventSource when headers are unavailable
      responses:
        '200':
          description: SSE stream (`text/event-stream`)
          content:
            text/event-stream:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: Too many SSE connections from this client IP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /v1/observability/import:
    post:
      tags: [observability]
      summary: Idempotent trace + agent + message ingest (MCPLab sync)
      operationId: importObservabilityTrace
      security:
        - ApiKeyAuth: []
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportTraceRequest'
      responses:
        '200':
          description: Import summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportTraceResponse'
        '400':
          $ref: '#/components/responses/ValidationFailed'
        '401':
          $ref: '#/components/responses/Unauthorized'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    BearerAuth:
      type: http
      scheme: bearer
    SseApiKeyQuery:
      type: apiKey
      in: query
      name: api_key

  responses:
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationFailed:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string

    RuntimeConfigResponse:
      type: object
      required: [ok, auth]
      properties:
        ok:
          type: boolean
          const: true
        auth:
          type: object
          required: [required, bearer, apiKeyHeader, sseQueryParam]
          properties:
            required:
              type: boolean
            bearer:
              type: boolean
              const: true
            apiKeyHeader:
              type: string
              const: x-api-key
            sseQueryParam:
              type: string
              const: api_key

    SnapshotResponse:
      type: object
      required: [ok, snapshot]
      properties:
        ok:
          type: boolean
          const: true
        snapshot:
          $ref: '#/components/schemas/PlaygroundSnapshot'

    PlaygroundSnapshot:
      type: object
      required: [server, agents, traces, trace_count, agent_links]
      properties:
        server:
          type: object
          required: [status, protocol_version, registered_agents, bus_open]
          properties:
            status:
              type: string
              enum: [healthy]
            protocol_version:
              type: string
            registered_agents:
              type: integer
              minimum: 0
            bus_open:
              type: boolean
        agents:
          type: array
          items:
            $ref: '#/components/schemas/AgentObservabilityRecord'
        traces:
          type: array
          items:
            $ref: '#/components/schemas/TraceListEntry'
        trace_count:
          type: integer
          minimum: 0
        active_trace:
          $ref: '#/components/schemas/TraceBundle'
        agent_links:
          type: array
          items:
            $ref: '#/components/schemas/AgentLink'

    AgentObservabilityRecord:
      type: object
      required: [id, name, version, capabilities, publicKey]
      properties:
        id:
          type: string
        name:
          type: string
        version:
          type: string
        capabilities:
          type: array
          items:
            type: string
        publicKey:
          oneOf:
            - type: string
            - type: object
        description:
          type: string
        metadata:
          type: object
          additionalProperties: true
        fleet:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [idle, active, error, offline]
        last_seen_at:
          type: string
          format: date-time
        active_trace_ids:
          type: array
          items:
            type: string
        message_count_24h:
          type: integer

    TraceListEntry:
      type: object
      required:
        - traceId
        - startedAt
        - lastActivityAt
        - messageCount
        - messageTypes
        - agents
      properties:
        traceId:
          type: string
        startedAt:
          type: string
          format: date-time
        lastActivityAt:
          type: string
          format: date-time
        messageCount:
          type: integer
        messageTypes:
          type: array
          items:
            type: string
        agents:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [running, completed, failed]
        completedAt:
          type: string
          format: date-time

    TraceBundle:
      type: object
      required:
        - trace_id
        - started_at
        - last_activity_at
        - message_count
        - message_types
        - agents
        - timeline
      properties:
        trace_id:
          type: string
        started_at:
          type: string
          format: date-time
        last_activity_at:
          type: string
          format: date-time
        message_count:
          type: integer
        message_types:
          type: array
          items:
            type: string
        agents:
          type: array
          items:
            type: string
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/TraceTimelineEvent'

    TraceTimelineEvent:
      type: object
      required: [index, timestamp, type, from, message_id, summary]
      properties:
        index:
          type: integer
        timestamp:
          type: string
          format: date-time
        type:
          type: string
        from:
          type: string
        to:
          type: string
        capability:
          type: string
        status:
          type: string
        message_id:
          type: string
        summary:
          type: string

    AgentLink:
      type: object
      required: [from_agent, to_agent, kind, message_count]
      properties:
        from_agent:
          type: string
        to_agent:
          type: string
        kind:
          type: string
        capability:
          type: string
        message_count:
          type: integer

    TraceGraphResponse:
      type: object
      required: [ok, graph]
      properties:
        ok:
          type: boolean
          const: true
        graph:
          $ref: '#/components/schemas/TraceGraphView'

    TraceGraphView:
      type: object
      required: [trace_id, layout, max_depth, nodes, edges]
      properties:
        trace_id:
          type: string
        layout:
          type: string
          enum: [hierarchical]
        max_depth:
          type: integer
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/TraceGraphNode'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/TraceGraphEdge'

    TraceGraphNode:
      type: object
      required: [agent_id, name, depth, status, capabilities]
      properties:
        agent_id:
          type: string
        name:
          type: string
        depth:
          type: integer
        fleet:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [idle, active, error, offline]
        capabilities:
          type: array
          items:
            type: string

    TraceGraphEdge:
      type: object
      required: [from_agent, to_agent, kind, message_count]
      properties:
        from_agent:
          type: string
        to_agent:
          type: string
        kind:
          type: string
        capability:
          type: string
        message_count:
          type: integer

    ImportTraceRequest:
      type: object
      required: [trace_id, messages]
      properties:
        trace_id:
          type: string
        agents:
          type: array
          items:
            type: object
            additionalProperties: true
        messages:
          type: array
          items:
            type: object
            additionalProperties: true
        completed_at:
          type: string
          format: date-time
        source:
          type: string

    ImportTraceResponse:
      type: object
      required: [ok, result]
      properties:
        ok:
          type: boolean
          const: true
        result:
          type: object
          required:
            - trace_id
            - imported_messages
            - skipped_messages
            - registered_agents
          properties:
            trace_id:
              type: string
            imported_messages:
              type: integer
            skipped_messages:
              type: integer
            registered_agents:
              type: integer